[core] fix controller and client types again
